requirements.txt
anthropic>=0.40.0
openai>=1.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
